refactor: working on splitting into multi files
Working on #2
